pyproject.toml
[tool.poetry]
name = "landhouse"
version = "0.1.0"
description = ""
authors = ["Akmaljon701 <akmalkhanever@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
django-cleanup = "^8.1.0"
django-silk = "^5.2.0"
django-modeltranslation = "^0.19.8"
pillow = "^10.4.0"
drf-spectacular-sidecar = "^2024.7.1"
psycopg2-binary = "^2.9.10"
psycopg = "^3.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
